+++ rust/operator/src/container.rs
@@ -9,15 +9,19 @@
 //! - Set resources
 //! - Add tcp probes and container ports (to the main containers)
 //!
-use crate::product_logging::{
-    FORMAT_NAMENODES_LOG4J_CONFIG_FILE, FORMAT_ZOOKEEPER_LOG4J_CONFIG_FILE, HDFS_LOG4J_CONFIG_FILE,
-    MAX_FORMAT_NAMENODE_LOG_FILE_SIZE, MAX_FORMAT_ZOOKEEPER_LOG_FILE_SIZE, MAX_HDFS_LOG_FILE_SIZE,
-    MAX_WAIT_NAMENODES_LOG_FILE_SIZE, MAX_ZKFC_LOG_FILE_SIZE, STACKABLE_LOG_DIR,
-    WAIT_FOR_NAMENODES_LOG4J_CONFIG_FILE, ZKFC_LOG4J_CONFIG_FILE,
+use crate::{
+    kerberos::principal_host_part,
+    product_logging::{
+        FORMAT_NAMENODES_LOG4J_CONFIG_FILE, FORMAT_ZOOKEEPER_LOG4J_CONFIG_FILE,
+        HDFS_LOG4J_CONFIG_FILE, MAX_FORMAT_NAMENODE_LOG_FILE_SIZE,
+        MAX_FORMAT_ZOOKEEPER_LOG_FILE_SIZE, MAX_HDFS_LOG_FILE_SIZE,
+        MAX_WAIT_NAMENODES_LOG_FILE_SIZE, MAX_ZKFC_LOG_FILE_SIZE, STACKABLE_LOG_DIR,
+        WAIT_FOR_NAMENODES_LOG4J_CONFIG_FILE, ZKFC_LOG4J_CONFIG_FILE,
+    },
 };
 
 use indoc::formatdoc;
-use snafu::{OptionExt, ResultExt, Snafu};
+use snafu::{ResultExt, Snafu};
 use stackable_hdfs_crd::{
     constants::{
         DATANODE_ROOT_DATA_DIR_PREFIX, DEFAULT_DATA_NODE_METRICS_PORT,
@@ -151,6 +155,8 @@ impl ContainerConfig {
     pub fn add_containers_and_volumes(
         pb: &mut PodBuilder,
         hdfs: &HdfsCluster,
+        hdfs_name: &str,
+        hdfs_namespace: &str,
         role: &HdfsRole,
         resolved_product_image: &ResolvedProductImage,
         merged_config: &(dyn MergedConfig + Send + 'static),
@@ -164,6 +170,8 @@ impl ContainerConfig {
         pb.add_volumes(main_container_config.volumes(merged_config, object_name));
         pb.add_container(main_container_config.main_container(
             hdfs,
+            hdfs_name,
+            hdfs_namespace,
             role,
             resolved_product_image,
             zk_config_map_name,
@@ -209,7 +217,7 @@ impl ContainerConfig {
                         SecretOperatorVolumeSourceBuilder::new(
                             &authentication_config.kerberos.secret_class,
                         )
-                        .with_service_scope(hdfs.name_any())
+                        .with_service_scope(hdfs_name)
                         .with_kerberos_service_name(role.kerberos_service_name())
                         .with_kerberos_service_name("HTTP")
                         .build(),
@@ -226,6 +234,8 @@ impl ContainerConfig {
                 pb.add_volumes(zkfc_container_config.volumes(merged_config, object_name));
                 pb.add_container(zkfc_container_config.main_container(
                     hdfs,
+                    hdfs_name,
+                    hdfs_namespace,
                     role,
                     resolved_product_image,
                     zk_config_map_name,
@@ -241,6 +251,8 @@ impl ContainerConfig {
                 );
                 pb.add_init_container(format_namenodes_container_config.init_container(
                     hdfs,
+                    hdfs_name,
+                    hdfs_namespace,
                     role,
                     resolved_product_image,
                     zk_config_map_name,
@@ -257,6 +269,8 @@ impl ContainerConfig {
                 );
                 pb.add_init_container(format_zookeeper_container_config.init_container(
                     hdfs,
+                    hdfs_name,
+                    hdfs_namespace,
                     role,
                     resolved_product_image,
                     zk_config_map_name,
@@ -274,6 +288,8 @@ impl ContainerConfig {
                 );
                 pb.add_init_container(wait_for_namenodes_container_config.init_container(
                     hdfs,
+                    hdfs_name,
+                    hdfs_namespace,
                     role,
                     resolved_product_image,
                     zk_config_map_name,
@@ -317,9 +333,13 @@ impl ContainerConfig {
     /// - Namenode ZooKeeper fail over controller (ZKFC)
     /// - Datanode main process
     /// - Journalnode main process
+
+    #[allow(clippy::too_many_arguments)]
     fn main_container(
         &self,
         hdfs: &HdfsCluster,
+        hdfs_name: &str,
+        hdfs_namespace: &str,
         role: &HdfsRole,
         resolved_product_image: &ResolvedProductImage,
         zookeeper_config_map_name: &str,
@@ -335,7 +355,7 @@ impl ContainerConfig {
 
         cb.image_from_product_image(resolved_product_image)
             .command(Self::command())
-            .args(self.args(hdfs, role, merged_config, &[])?)
+            .args(self.args(hdfs, hdfs_name, hdfs_namespace, role, merged_config, &[]))
             .add_env_vars(self.env(
                 hdfs,
                 zookeeper_config_map_name,
@@ -364,6 +384,8 @@ impl ContainerConfig {
     fn init_container(
         &self,
         hdfs: &HdfsCluster,
+        hdfs_name: &str,
+        hdfs_namespace: &str,
         role: &HdfsRole,
         resolved_product_image: &ResolvedProductImage,
         zookeeper_config_map_name: &str,
@@ -376,7 +398,14 @@ impl ContainerConfig {
 
         cb.image_from_product_image(resolved_product_image)
             .command(Self::command())
-            .args(self.args(hdfs, role, merged_config, namenode_podrefs)?)
+            .args(self.args(
+                hdfs,
+                hdfs_name,
+                hdfs_namespace,
+                role,
+                merged_config,
+                namenode_podrefs,
+            ))
             .add_env_vars(self.env(hdfs, zookeeper_config_map_name, env_overrides, None))
             .add_volume_mounts(self.volume_mounts(hdfs, merged_config));
 
@@ -427,10 +456,12 @@ impl ContainerConfig {
     fn args(
         &self,
         hdfs: &HdfsCluster,
+        hdfs_name: &str,
+        hdfs_namespace: &str,
         role: &HdfsRole,
         merged_config: &(dyn MergedConfig + Send + 'static),
         namenode_podrefs: &[HdfsPodRef],
-    ) -> Result<Vec<String>, Error> {
+    ) -> Vec<String> {
         let mut args = String::new();
         args.push_str(&self.create_config_directory_cmd());
         args.push_str(&self.copy_config_xml_cmd());
@@ -489,7 +520,7 @@ wait_for_termination $!
                 // If there is no active namenode, the current pod is not formatted we format as
                 // active namenode. Otherwise as standby node.
                 if hdfs.has_kerberos_enabled() {
-                    args.push_str(&Self::get_kerberos_ticket(hdfs, role)?);
+                    args.push_str(&Self::get_kerberos_ticket(hdfs_name, hdfs_namespace, role));
                 }
                 args.push_str(&formatdoc!(
                     r###"
@@ -570,7 +601,7 @@ wait_for_termination $!
                     ));
                 }
                 if hdfs.has_kerberos_enabled() {
-                    args.push_str(&Self::get_kerberos_ticket(hdfs, role)?);
+                    args.push_str(&Self::get_kerberos_ticket(hdfs_name, hdfs_namespace, role));
                 }
                 args.push_str(&formatdoc!(
                     r###"
@@ -610,7 +641,7 @@ wait_for_termination $!
                 ));
             }
         }
-        Ok(vec![args])
+        vec![args]
     }
 
     // Command to export `KERBEROS_REALM` env var to default real from krb5.conf, e.g. `CLUSTER.LOCAL`
@@ -622,19 +653,18 @@ wait_for_termination $!
     /// Command to `kinit` a ticket using the principal created for the specified hdfs role
     /// Needs the KERBEROS_REALM env var, which will be written with `export_kerberos_real_env_var_command`
     /// Needs the POD_NAME env var to be present, which will be provided by the PodSpec
-    fn get_kerberos_ticket(hdfs: &HdfsCluster, role: &HdfsRole) -> Result<String, Error> {
+    fn get_kerberos_ticket(hdfs_name: &str, hdfs_namespace: &str, role: &HdfsRole) -> String {
         let principal = format!(
-            "{service_name}/{hdfs_name}.{namespace}.svc.cluster.local@${{KERBEROS_REALM}}",
+            "{service_name}/{principal_host_part}",
             service_name = role.kerberos_service_name(),
-            hdfs_name = hdfs.name_any(),
-            namespace = hdfs.namespace().context(ObjectHasNoNamespaceSnafu)?,
+            principal_host_part = principal_host_part(hdfs_name, hdfs_namespace),
         );
-        Ok(formatdoc!(
+        formatdoc!(
             r###"
             echo "Getting ticket for {principal}" from /stackable/kerberos/keytab
             kinit "{principal}" -kt /stackable/kerberos/keytab
             "###,
-        ))
+        )
     }
 
     fn get_namenode_service_state_command() -> String {